Chance Argument and No longer using |
The new chance argument now controls the amount of characters that get replaces, and can range from 00% to 99%. The percent sign is optional as it gets replaces durring the conversion process.

The pipe symbol is no longer being used, and all arguments after the chance argument are being combined into one string that then gets used as the message. This allows the user to have more flexibility, and makes the command look nicer.