Add AutoComplete (WIP)
This is still unfinished, but it's pretty stable and usable at this
point.

It intentionally has no settings yet, because I might still make changes
at this point that would completely change the way it works. Such as:
- Adding support for OpenAI's ChatGPT models, which need a completely
different prompt
- Adding support for other providers like GooseAI, which have a
completely different API with different parameters, etc.
- Adding support for self-hosted language models like LLaMA-7B, which
have very different parameters again

We might also need some new setting types for this hack, such as a
StringSetting for any custom prompts and an EncryptedStringSetting to
store the API key. The current method of storing the API key in an
environment variable works, but it's not very user-friendly.