Fix relHitVec calculation
The previous code assumed that the center of a block is the same as its
size divided by two, which works for most blocks but not all of them.
For example, an upside down slab has its center on the Y axis at 0.75,
but half of its height is only 0.25.

Breaking a slab from the top with the old code would place relHitVec's Y
coordinate at 0.75+0.75=1.5 (floating half a block above the slab's top
surface), with the new code it's 0.75+0.25=1 (exactly on the surface
where it should be).