Move player null check to HandleInputEvent mixin
After 0d50ef6e41db6640862795511147ede4a8d217e3, it was possible for this
event to be called outside of gameplay, which the hacks using it are not
designed for.

This commit extends the fix from #1037 to also cover Killaura and
KillauraLegit, which could still sometimes crash if they were attacking
while leaving a world or server.