"Fix" LSD shader
This still won't work because Mojang broke the matrix4x4 uniform type
(listed as matrix4 in the changelog, but matrix4x4 is correct). It needs
16 inputs but the number of inputs per uniform is limited to 4. Removing
that limit with a mixin causes the game to crash, so this seems to be a
more complex design flaw in the 25w07a shader change. Minecraft itself
no longer uses this type anywhere, so it likely went unnoticed.